refactor(service): split sync.rs tests into a submodule
